+++ src/MinFraud/Validation/Rules/Event.php
@@ -23,6 +23,7 @@ class Event extends AbstractWrapper
                         'account_login',
                         'email_change',
                         'password_reset',
+                        'payout_change',
                         'purchase',
                         'recurring_purchase',
                         'referral',